+++ nitrocli/CHANGELOG.md
@@ -3,6 +3,8 @@ Unreleased
 - Added the `reset` command to perform a factory reset
 - Added the `-V`/`--version` option to print the program's version
 - Check the status of a PWS slot before accessing it in `pws get`
+- Added `NITROCLI_NO_CACHE` environment variable to bypass caching of
+  secrets
 - Bumped `libc` dependency to `0.2.55`
 - Bumped `cc` dependency to `1.0.37`